fix pre-commit failure because of missing submodule
