--- projects/rocprofiler-compute/src/utils/gpu_arch.py
@@ -1,17 +1,0 @@
-# Copyright (c) Advanced Micro Devices, Inc.
-# SPDX-License-Identifier:  MIT
-
-"""Lightweight GPU architecture canonicalization helpers."""
-
-from __future__ import annotations
-
-from typing import Optional
-
-
-def canonical_config_arch(gpu_arch: Optional[str]) -> Optional[str]:
-    """Map GPU architectures to the shared analysis-config directory name."""
-    if gpu_arch is None:
-        return None
-    if gpu_arch.startswith("gfx115"):
-        return "gfx115x"
-    return gpu_arch
