+++ projects/rocprofiler-compute/tests/test_gpu_specs.py
@@ -17,14 +17,14 @@ try:
     import src.utils.specs as specs
     from src.rocprof_compute_soc.soc_base import OmniSoC_Base
     from src.utils.file_io import is_single_panel_config
-    from src.utils.gpu_arch import canonical_config_arch
     from src.utils.specs import generate_machine_specs
+    from src.utils.utils_common import canonical_config_arch
 except Exception:
     import utils.specs as specs
     from rocprof_compute_soc.soc_base import OmniSoC_Base
     from utils.file_io import is_single_panel_config
-    from utils.gpu_arch import canonical_config_arch
     from utils.specs import generate_machine_specs
+    from utils.utils_common import canonical_config_arch
 
 # NOTE: Only testing gfx942 for now.
 GFX942_CHIP_IDS_TO_NUM_XCDS = {
